Fix std::accumulate overflow

+++ mindspore/ops/kernel/ascend/aclnn/pyboost_impl/customize/group_norm.cc
@@ -37,8 +37,9 @@ void GroupNormAscendCustomize(const std::shared_ptr<OpRunner> &op, const TensorP
   const auto &shape = input_tensor->shape();
   const int64_t N = shape[0];
   const int64_t C = shape[1];
-  const int64_t HxW =
-    (shape.size() == kDim2) ? 1 : std::accumulate(shape.begin() + 2, shape.end(), 1, std::multiplies<int64_t>());
+  const int64_t HxW = (shape.size() == kDim2) ? 1
+                                              : std::accumulate(shape.begin() + 2, shape.end(), static_cast<int64_t>(1),
+                                                                std::multiplies<int64_t>());
   PyBoostUtils::PrepareOpInputs(op->device_context(), op->stream_id(), input_tensor, gamma_opt_tensor, beta_opt_tensor);
   PyBoostUtils::PrepareOpOutputs(op->device_context(), op->stream_id(), op->outputs());
   // Async
