+++ mindspore/ops/kernel/ascend/aicpu/aicpu_ops/cpu_kernel/ms_kernel/random_shuffle.cc
@@ -79,7 +79,8 @@ uint32_t RandomShuffleCpuKernel::RandomShuffleCompute(CpuKernelContext &ctx, Ten
   }
 
   size_t batch_size = input_shape[0];
-  size_t block_num = std::accumulate(input_shape.begin() + 1, input_shape.end(), 1, std::multiplies<int64_t>());
+  size_t block_num =
+    std::accumulate(input_shape.begin() + 1, input_shape.end(), static_cast<int64_t>(1), std::multiplies<int64_t>());
   size_t block_size = GetSizeByDataType(input->GetDataType()) * block_num;
 
   std::vector<int64_t> shuffled_indices(batch_size);
