+++ mindspore/ops/kernel/gpu/cuda/arrays/mvlgamma_gpu_kernel.cc
@@ -54,7 +54,8 @@ int MvlgammaGpuKernelMod::Resize(const std::vector<KernelTensor *> &inputs,
   ResetResource();
   std::vector<int64_t> input_shape = std::vector<int64_t>(inputs.at(kIndex0)->GetDeviceShapeVector().begin(),
                                                           inputs.at(kIndex0)->GetDeviceShapeVector().end());
-  input_elements_ = std::accumulate(input_shape.begin(), input_shape.end(), 1, std::multiplies<int64_t>());
+  input_elements_ =
+    std::accumulate(input_shape.begin(), input_shape.end(), static_cast<int64_t>(1), std::multiplies<int64_t>());
   if (input_elements_ == 0) {
     is_null_input_ = true;
   }
