+++ mindspore/ops/kernel/gpu/cuda/sparse/sparse_segment_ops_gpu_kernel.cc
@@ -70,7 +70,8 @@ int SparseSegmentOpsGpuKernelMod::Resize(const std::vector<KernelTensor *> &inpu
   }
   ResetResource();
   std::vector<int64_t> output_shape = outputs.at(kIndex0)->GetShapeVector();
-  output_elements_ = std::accumulate(output_shape.begin(), output_shape.end(), 1, std::multiplies<int64_t>());
+  output_elements_ =
+    std::accumulate(output_shape.begin(), output_shape.end(), static_cast<int64_t>(1), std::multiplies<int64_t>());
   if (output_elements_ == 0) {
     is_null_input_ = true;
   }
